Move generators into ViewComponent namespace
