+++ lib/generators/view_component/abstract_generator.rb
