+++ lib/generators/view_component/component/USAGE
