+++ lib/generators/view_component/component/templates/component.rb.tt
