+++ lib/generators/view_component/erb/templates/component.html.erb.tt
