+++ lib/generators/view_component/haml/haml_generator.rb
@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
-require "rails/generators/erb/component_generator"
+require "generators/view_component/erb/erb_generator"
 
-module Slim
+module ViewComponent
   module Generators
-    class ComponentGenerator < Erb::Generators::ComponentGenerator
+    class HamlGenerator < ViewComponent::Generators::ErbGenerator
       include ViewComponent::AbstractGenerator
 
       source_root File.expand_path("templates", __dir__)
       class_option :sidecar, type: :boolean, default: false
 
       def engine_name
-        "slim"
+        "haml"
       end
 
       def copy_view_file
