+++ lib/generators/view_component/haml/templates/component.html.haml.tt
