+++ lib/generators/view_component/locale/locale_generator.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require "rails/generators/abstract_generator"
+require "generators/view_component/abstract_generator"
 
-module Locale
+module ViewComponent
   module Generators
-    class ComponentGenerator < ::Rails::Generators::NamedBase
+    class LocaleGenerator < ::Rails::Generators::NamedBase
       include ViewComponent::AbstractGenerator
 
       source_root File.expand_path("templates", __dir__)
