+++ lib/generators/view_component/preview/templates/component_preview.rb.tt
