+++ lib/generators/view_component/rspec/templates/component_spec.rb.tt
