+++ lib/generators/view_component/slim/templates/component.html.slim.tt
