+++ lib/generators/view_component/stimulus/templates/component_controller.js.tt
