+++ lib/generators/view_component/tailwindcss/templates/component.html.erb.tt
