+++ lib/generators/view_component/test_unit/templates/component_test.rb.tt
