--- lib/rails/generators/tailwindcss/component_generator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require "rails/generators/erb/component_generator"
-
-module Tailwindcss
-  module Generators
-    class ComponentGenerator < Erb::Generators::ComponentGenerator
-      source_root File.expand_path("templates", __dir__)
-    end
-  end
-end
